+++ src/main/resources/com/dataintuitive/viash/helpers/bashutils/ViashDockerFuns.sh
@@ -64,26 +64,28 @@ function ViashDockerPullElseBuild {
 # $2                  : docker setup strategy, see DockerSetupStrategy.scala
 # ViashDockerBuild    : a Bash function which builds a docker image, takes image identifier as argument.
 # examples:
-#   ViashDockerPullElseBuild mynewcomponent alwaysbuild
+#   ViashDockerSetup mynewcomponent alwaysbuild
 function ViashDockerSetup {
   VSHD_ID="$1"
   VSHD_STRAT="$2"
-  if [ "$VSHD_STRAT" == "alwaysbuild" -o "$VSHD_STRAT" == "build" ]; then
+  if [ "$VSHD_STRAT" == "alwaysbuild" -o "$VSHD_STRAT" == "build" -o "$VSHD_STRAT" == "b" ]; then
     ViashDockerBuild $VSHD_ID --no-cache
-  elif [ "$VSHD_STRAT" == "alwayspull" -o "$VSHD_STRAT" == "pull" ]; then
+  elif [ "$VSHD_STRAT" == "alwayspull" -o "$VSHD_STRAT" == "pull" -o "$VSHD_STRAT" == "p" ]; then
     ViashDockerPull $VSHD_ID
   elif [ "$VSHD_STRAT" == "alwayspullelsebuild" -o "$VSHD_STRAT" == "pullelsebuild" ]; then
     ViashDockerPullElseBuild $VSHD_ID --no-cache
   elif [ "$VSHD_STRAT" == "alwayspullelsecachedbuild" -o "$VSHD_STRAT" == "pullelsecachedbuild" ]; then
     ViashDockerPullElseBuild $VSHD_ID
-  elif [ "$VSHD_STRAT" == "alwayscachedbuild" -o "$VSHD_STRAT" == "cachedbuild" ]; then
+  elif [ "$VSHD_STRAT" == "alwayscachedbuild" -o "$VSHD_STRAT" == "cachedbuild" -o "$VSHD_STRAT" == "cb" ]; then
     ViashDockerBuild $VSHD_ID
-  elif [ "$VSHD_STRAT" == "donothing" -o "$VSHD_STRAT" == "meh" ]; then
-    echo "Skipping setup."
   elif [[ "$VSHD_STRAT" =~ ^ifneedbe ]]; then
+    set +e
     ViashDockerLocalTagCheck $VSHD_ID
-    if [ $? -eq 0 ]; then
-      echo "Image $VSHD_ID already exists"
+    outCheck=$?
+    set -e
+    if [ $outCheck -eq 0 ]; then
+      # echo "Image $VSHD_ID already exists"
+      :
     elif [ "$VSHD_STRAT" == "ifneedbebuild" ]; then
       ViashDockerBuild $VSHD_ID --no-cache
     elif [ "$VSHD_STRAT" == "ifneedbecachedbuild" ]; then
@@ -96,25 +98,11 @@ function ViashDockerSetup {
       ViashDockerPullElseBuild $VSHD_ID
     else
       echo "Unrecognised Docker strategy: $VSHD_STRAT"
+      exit 1
     fi
-  else
-    echo "Unrecognised Docker strategy: $VSHD_STRAT"
-  fi
-}
-
-# ViashDockerPush: create a Docker image, according to specified docker setup strategy
-#
-# $1                  : image identifier with format `[registry/]image[:tag]`
-# $2                  : docker setup strategy, see DockerPushStrategy.scala
-# ViashDockerBuild    : a Bash function which builds a docker image, takes image identifier as argument.
-# examples:
-#   ViashDockerPullElseBuild mynewcomponent alwaysbuild
-function ViashDockerPush {
-  VSHD_ID="$1"
-  VSHD_STRAT="$2"
-  if [ "$VSHD_STRAT" == "alwayspush" -o "$VSHD_STRAT" == "force" ]; then
+  elif [ "$VSHD_STRAT" == "push" -o "$VSHD_STRAT" == "forcepush" -o "$VSHD_STRAT" == "alwayspush" ]; then
     set +e
-    docker push $1
+    docker push $VSHD_ID
     outPush=$?
     set -e
     if [ $outPush -eq 0 ]; then
@@ -123,9 +111,9 @@ function ViashDockerPush {
       echo "> $VSHD_ID force push ... error"
       exit 1
     fi
-  elif [ "$VSHD_STRAT" == "pushifnotpresent" ]; then
+  elif [ "$VSHD_STRAT" == "pushifnotpresent" -o "$VSHD_STRAT" == "gentlepush" -o "$VSHD_STRAT" == "maybepush" ]; then
     set +e
-    ViashDockerRemoteTagCheck $1
+    ViashDockerRemoteTagCheck $VSHD_ID
     outCheck=$?
     set -e
     if [ $outCheck -eq 0 ]; then
@@ -140,11 +128,16 @@ function ViashDockerPush {
         echo "... ok"
       else
         echo "... error"
+        exit 1
       fi
     fi
+  elif [ "$VSHD_STRAT" == "donothing" -o "$VSHD_STRAT" == "meh" ]; then
+    echo "Skipping setup."
   else
-    echo "Unrecognised Docker push strategy: $VSHD_STRAT"
+    echo "Unrecognised Docker strategy: $VSHD_STRAT"
+    exit 1
   fi
 }
+
 
 ######## End of helper functions for setting up Docker images for viash ########