+++ src/main/scala/com/dataintuitive/viash/platforms/DockerPlatform.scala
@@ -39,8 +39,7 @@ case class DockerPlatform(
   chown: Boolean = true,
   port: Option[List[String]] = None,
   workdir: Option[String] = None,
-  setup_strategy: DockerSetupStrategy = AlwaysCachedBuild,
-  push_strategy: DockerPushStrategy = PushIfNotPresent,
+  setup_strategy: DockerSetupStrategy = IfNeedBePullElseCachedBuild,
   privileged: Boolean = false,
 
   // setup variables
@@ -86,7 +85,7 @@ case class DockerPlatform(
     }
 
     // create setup
-    val (effectiveID, setupCommands, setupMods) = processDockerSetup(functionality)
+    val (effectiveID, setupMods) = processDockerSetup(functionality)
 
     // generate automount code
     val dmVol = processDockerVolumes(functionality)
@@ -120,7 +119,6 @@ case class DockerPlatform(
       text = Some(BashWrapper.wrapScript(
         executor = executor,
         functionality = fun2,
-        setupCommands = setupCommands,
         mods = dm
       ))
     )
@@ -199,87 +197,63 @@ case class DockerPlatform(
              |  out=$$?
              |  set -e
              |  if [ ! $$out -eq 0 ]; then
-             |    echo "> ERROR: Something went wrong building the container $$@"
+             |    echo "> ERROR: Something went wrong while building the container $$@"
              |    echo "> Error transcript follows:"
-             |    cat $$tmpdir/docker_build.log
+             |    cat "$$tmpdir/docker_build.log"
              |    echo "> --- end of error transcript"
-             |  fi
-             |  if [ "$$VIASH_EXEC_MODE" == "setup" ]; then
-             |    exit $$out
              |  fi""".stripMargin
 
         (vdf, vdb)
       }
 
-    val setupCommands = {
-      s"""# ViashDockerFile: print the dockerfile to stdout
+    val preParse =
+      s"""
+         |${Bash.ViashDockerFuns}
+         |
+         |# ViashDockerFile: print the dockerfile to stdout
          |# return : dockerfile required to run this component
          |# examples:
          |#   ViashDockerFile
          |function ViashDockerfile {
          |$viashDockerFile
          |}
-         |# ViashDockerBuild: ...
+         |
+         |# ViashDockerBuild: build a docker container
+         |# $$1              : image identifier with format `[registry/]image[:tag]`
+         |# exit code $$?    : whether or not the image was built
          |function ViashDockerBuild {
          |$viashDockerBuild
-         |}
-         |
-         |# ViashSetup: ...
-         |function ViashSetup {
-         |  ViashDockerSetup $effectiveID $$$dockerSetupStrategyVar
-         |}
-         |
-         |# ViashPush: ...
-         |function ViashPush {
-         |  ViashDockerPush $effectiveID $$$dockerPushStrategyVar
          |}""".stripMargin
-    }
-
-    val preParse =
-      s"""
-         |${Bash.ViashDockerFuns}
-         |# initialise variables
-         |$dockerSetupStrategyVar='${setup_strategy.id}'
-         |$dockerPushStrategyVar='${push_strategy.id}'""".stripMargin
 
     val parsers =
       s"""
-         |        ---dss|---docker_setup_strategy)
-         |            ${BashWrapper.var_exec_mode}="setup"
-         |            $dockerSetupStrategyVar="$$2"
-         |            shift 2
+         |        ---setup)
+         |            ViashDockerSetup '$effectiveID' "$$2"
+         |            exit 0
          |            ;;
-         |        ---docker_setup_strategy=*)
-         |            ${BashWrapper.var_exec_mode}="setup"
-         |            $dockerSetupStrategyVar=$$(ViashRemoveFlags "$$2")
-         |            shift 1
-         |            ;;
-         |        ---dps|---docker_push_strategy)
-         |            ${BashWrapper.var_exec_mode}="push"
-         |            $dockerPushStrategyVar="$$2"
-         |            shift 2
-         |            ;;
-         |        ---docker_push_strategy=*)
-         |            ${BashWrapper.var_exec_mode}="push"
-         |            $dockerPushStrategyVar=$$(ViashRemoveFlags "$$2")
-         |            shift 1
+         |        ---setup=*)
+         |            ViashDockerSetup '$effectiveID' "$$(ViashRemoveFlags "$$1")"
+         |            exit 0
          |            ;;
          |        ---dockerfile)
          |            ViashDockerfile
          |            exit 0
          |            ;;""".stripMargin
 
+    def postParse =
+      s"""
+         |ViashDockerSetup '$effectiveID' ${IfNeedBePullElseCachedBuild.id}""".stripMargin
+
     val mods = BashWrapperMods(
       preParse = preParse,
-      parsers = parsers
+      parsers = parsers,
+      postParse = postParse
     )
 
-    (effectiveID, setupCommands, mods)
+    (effectiveID, mods)
   }
 
   private val extraMountsVar = "VIASH_EXTRA_MOUNTS"
-  private val dockerSetupStrategyVar = "VIASH_DOCKER_SETUP_STRATEGY"
-  private val dockerPushStrategyVar = "VIASH_DOCKER_PUSH_STRATEGY"
 
   private def processDockerVolumes(functionality: Functionality) = {
     val args = functionality.argumentsAndDummies