+++ src/main/scala/com/dataintuitive/viash/platforms/NativePlatform.scala
@@ -43,32 +43,12 @@ case class NativePlatform(
       dest = Some(functionality.name),
       text = Some(BashWrapper.wrapScript(
         executor = executor,
-        functionality = functionality,
-        setupCommands = setupCommands
+        functionality = functionality
       ))
     )
 
     functionality.copy(
       resources = Some(bashScript :: functionality.resources.getOrElse(Nil).tail)
     )
-  }
-
-  def setupCommands: String = {
-    val runCommands = requirements.map(_.installCommands)
-
-    val commands =
-      runCommands.map(li =>
-        if (li.isEmpty) {
-          ""
-        } else {
-          "cat << 'HERE'\n" +
-          "# run the following commands:\n" +
-          li.mkString("", " && \\\n  ", "\n") +
-          "HERE\n"
-        }
-      ).mkString
-
-    s"""function ViashSetup {
-       |${if (commands == "") ":\n" else commands}}""".stripMargin
   }
 }